# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.ArrayResult:
    properties:
      count:
        type: integer
      pages:
        type: integer
    type: object
  model.PersonData:
    properties:
      address:
        type: string
      age:
        type: integer
      id:
        type: integer
      name:
        type: string
      work:
        type: string
    type: object
  model.PersonRequest:
    properties:
      address:
        type: string
      age:
        type: integer
      name:
        type: string
      work:
        type: string
    type: object
host: rsoi.rigellab.ru
info:
  contact:
    name: RigelLab
    url: https://t.me/rigellab
  description: Апи для лабы 1
  title: Persons API
  version: "0.1"
paths:
  /persons:
    get:
      consumes:
      - application/json
      parameters:
      - description: Поиск по имени, адресу, работе
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Пользователи
          schema:
            allOf:
            - $ref: '#/definitions/model.ArrayResult'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/model.PersonData'
                  type: array
              type: object
      summary: Поиск пользователей
      tags:
      - persons
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные о пользователе
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PersonRequest'
      responses:
        "201":
          description: Пользователь создан
        "400":
          description: Ошибка в теле запроса
      summary: Создание пользователя
      tags:
      - persons
  /persons/{id}:
    delete:
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Пользователь удалён
        "404":
          description: Пользователь не найден
      summary: Удаление пользователя
      tags:
      - persons
    get:
      consumes:
      - application/json
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь
          schema:
            $ref: '#/definitions/model.PersonData'
        "404":
          description: Пользователь не найден
      summary: Пользователь
      tags:
      - persons
    patch:
      consumes:
      - application/json
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные о пользователе
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PersonRequest'
      responses:
        "204":
          description: Пользователь обновлён
        "400":
          description: Ошибка в теле запроса
        "404":
          description: Пользователь не найден
      summary: Обновление пользователя
      tags:
      - persons
swagger: "2.0"